Add decompose, decomposeLast and error example
This is for #1. I wonder if it would make sense to add some of this
nontrivial machinery for joining simple *-kinded types into a sum for
things like error handling into a module or a separate
library. There's nothing that inherently ties this to error
handling. It could just as easily be any other kind of sum type, but
sometimes you don't need all the functor features of Data.Sum and I
think the API here is pretty nice to work with.